ipcamera debug
