+++ modules/drivers/ipcamera/BUILD
@@ -3,24 +3,74 @@ load("//tools:cpplint.bzl", "cpplint")
 
 package(default_visibility = ["//visibility:public"])
 
+#cc_binary(
+#    name = "ipcamera_component.so",
+#    linkshared = True,
+#    linkstatic = False,
+#    deps = [":ipcamera_component_lib"],
+#)
+#
+#cc_library(
+#    name = "ipcamera_component_lib",
+#    srcs = ["ipcamera_component.cc"],
+#    hdrs = ["ipcamera_component.h",
+#            "ipcamera_main.h"],
+#    deps = [
+#        "//cyber",
+#        "//modules/drivers/proto:sensor_image_cc_proto",
+#        "//modules/drivers/ipcamera/proto:ipcamera_conf_cc_proto",
+#        "@ipcamera",
+#        "@opencv",
+#    ],
+#)
+
 cc_binary(
     name = "ipcamera_component.so",
     linkshared = True,
     linkstatic = False,
     deps = [":ipcamera_component_lib"],
 )
-
 cc_library(
     name = "ipcamera_component_lib",
     srcs = ["ipcamera_component.cc"],
-    hdrs = ["ipcamera_component.h",
-            "ipcamera_main.h"],
+    hdrs = ["ipcamera_component.h"],
+    copts = ["-DMODULE_NAME=\\\"ipcamera\\\""],
     deps = [
+        ":ipcamera_main_lib",
         "//cyber",
-        "//cyber/examples/proto:examples_cc_proto",
         "//modules/drivers/proto:sensor_image_cc_proto",
         "//modules/drivers/ipcamera/proto:ipcamera_conf_cc_proto",
         "@ipcamera",
+        "@opencv",
+    ],
+)
+cc_library(
+    name = "ipcamera_main_lib",
+    srcs = ["ipcamera_main.cc"],
+    hdrs = ["ipcamera_main.h"],
+    deps = [
+        #"//cyber",
+        "@ipcamera",
+    ],
+)
+cc_binary(
+    name = "ipcamera_reader.so",
+    linkopts = ["-shared"],
+    linkshared = True,
+    linkstatic = False,
+    deps = [
+        ":ipcamera_reader_lib",
+    ],
+)
+
+cc_library(
+    name = "ipcamera_reader_lib",
+    srcs = ["ipcamera_reader.cc"],
+    hdrs = ["ipcamera_reader.h"],
+    copts = ["-DMODULE_NAME=\\\"ipcamera_reader\\\""],
+    deps = [
+        "//cyber",
+        "//modules/drivers/proto:sensor_image_cc_proto",
         "@opencv",
     ],
 )
